updated style.css, index.html, and script.js
I fixed the resume button and added box shadow to my cards. In my index file, I added the code and life buttons as well as a second header for my description. I fixed the image value in my object and made it so that my project cards are added depending on my JavaScript object